Add inventory system
Forgot to commit from Friday oops.

Quest system creates the quest with stats, and the checker pulls both the quest stats and adventurer stats to compare.

Reward is created at time of quest creation and displays potential rewards.

If successful that item is placed in adventurer's inventory, if fail they come home.